refactor(core): move up config.plugin
